Consolidate TranslateExpressions into RowExpressionRewriteRuleSet
